# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AvatarLook {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatar_type?: string | null;
  /** @nullable */
  preview_image_url?: string | null;
  /** @nullable */
  preview_video_url?: string | null;
  /** @nullable */
  gender?: string | null;
}

export interface AvatarLookList {
  avatars: AvatarLook[];
  /** @nullable */
  next_token?: string | null;
}

export interface Voice {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  preview_audio_url?: string | null;
}

export interface VoiceList {
  voices: Voice[];
  /** @nullable */
  next_token?: string | null;
}

export interface Video {
  id: string;
  status: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  created_at?: number | null;
  /** @nullable */
  error?: string | null;
}

export interface VideoList {
  videos: Video[];
  /** @nullable */
  next_token?: string | null;
}

export type VideoInputOrientation = typeof VideoInputOrientation[keyof typeof VideoInputOrientation];


export const VideoInputOrientation = {
  landscape: 'landscape',
  portrait: 'portrait',
} as const;

export interface VideoInput {
  avatar_id: string;
  voice_id: string;
  script: string;
  /** @nullable */
  title?: string | null;
  orientation?: VideoInputOrientation;
}

export type VideoAgentInputOrientation = typeof VideoAgentInputOrientation[keyof typeof VideoAgentInputOrientation];


export const VideoAgentInputOrientation = {
  landscape: 'landscape',
  portrait: 'portrait',
} as const;

export interface VideoAgentInput {
  prompt: string;
  /** @nullable */
  avatar_id?: string | null;
  /** @nullable */
  voice_id?: string | null;
  orientation?: VideoAgentInputOrientation;
}

export interface VideoCreateResult {
  video_id: string;
  /** @nullable */
  session_id?: string | null;
  status: string;
}

export interface Credits {
  /** @nullable */
  remaining_credits?: number | null;
  /** @nullable */
  balance?: number | null;
  /** @nullable */
  credit_type?: string | null;
}

export type ListAvatarLooksParams = {
limit?: number;
/**
 * @nullable
 */
token?: string | null;
/**
 * @nullable
 */
ownership?: string | null;
};

export type ListVoicesParams = {
limit?: number;
/**
 * @nullable
 */
token?: string | null;
/**
 * @nullable
 */
language?: string | null;
/**
 * @nullable
 */
gender?: string | null;
};

export type ListVideosParams = {
limit?: number;
/**
 * @nullable
 */
token?: string | null;
};