# lib/api-zod/src/generated/types/avatarLook.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AvatarLook {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatar_type?: string | null;
  /** @nullable */
  preview_image_url?: string | null;
  /** @nullable */
  preview_video_url?: string | null;
  /** @nullable */
  gender?: string | null;
}