# lib/api-zod/src/generated/types/avatarLookList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { AvatarLook } from './avatarLook';

export interface AvatarLookList {
  avatars: AvatarLook[];
  /** @nullable */
  next_token?: string | null;
}