# lib/api-zod/src/generated/types/credits.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Credits {
  /** @nullable */
  remaining_credits?: number | null;
  /** @nullable */
  balance?: number | null;
  /** @nullable */
  credit_type?: string | null;
}