# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: string;
  status: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  thumbnail_url?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  created_at?: number | null;
  /** @nullable */
  error?: string | null;
}