# lib/api-zod/src/generated/types/videoAgentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoAgentInputOrientation } from './videoAgentInputOrientation';

export interface VideoAgentInput {
  prompt: string;
  /** @nullable */
  avatar_id?: string | null;
  /** @nullable */
  voice_id?: string | null;
  orientation?: VideoAgentInputOrientation;
}