# lib/api-zod/src/generated/types/videoCreateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoCreateResult {
  video_id: string;
  /** @nullable */
  session_id?: string | null;
  status: string;
}