# lib/api-zod/src/generated/types/videoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoInputOrientation } from './videoInputOrientation';

export interface VideoInput {
  avatar_id: string;
  voice_id: string;
  script: string;
  /** @nullable */
  title?: string | null;
  orientation?: VideoInputOrientation;
}