# lib/api-zod/src/generated/types/videoList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Video } from './video';

export interface VideoList {
  videos: Video[];
  /** @nullable */
  next_token?: string | null;
}