# lib/api-zod/src/generated/types/voice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Voice {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  preview_audio_url?: string | null;
}