# lib/api-zod/src/generated/types/voiceList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HeyGen Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Voice } from './voice';

export interface VoiceList {
  voices: Voice[];
  /** @nullable */
  next_token?: string | null;
}